requirements.txt
autopep8==2.0.0
passlib==1.7.4
pycodestyle==2.9.1
pycryptodome==3.15.0
tomli==2.0.1
